+++ broker-service/internal/event/emitter.go
