+++ broker-service/internal/logs/logs.proto
